create abstract base class for aem upload classes.
