+++ lib/aem/aembinaryupload.js
@@ -16,30 +16,14 @@ require("core-js/stable");
 
 const fileUrl = require("file-url");
 
-const { AssetMetadata } = require("../asset/assetmetadata");
-const { NameConflictPolicy } = require("../asset/nameconflictpolicy");
-const { AEMInitiateUpload } = require("../functions/aeminitiateupload");
 const { IllegalArgumentError } = require("../error");
 const { Asset } = require("../asset/asset");
-const { TransferController, TransferEvents } = require("../controller/transfercontroller");
 const { TransferAsset } = require("../asset/transferasset");
-const { AEMCompleteUpload } = require("../functions/aemcompleteupload");
-const { CreateTransferParts } = require("../functions/transferpartscreate");
-const { JoinTransferParts } = require("../functions/transferpartsjoin");
-const { FailUnsupportedAssets } = require("../functions/failunsupportedassets");
-const { CloseFiles } = require("../functions/closefiles");
-const { MapConcurrent } = require("../generator/mapconcurrent");
-const { Transfer } = require("../functions/transfer");
-
-const ErrorCodes = require("../http-error-codes");
 
 const PRIVATE = Symbol("PRIVATE");
 
-const MAX_RETRY_COUNT = 5;
-
 /**
- * A transfer process that uses AEM's direct binary upload algorithm to upload assets
- * to an AEM instance.
+ * A transfer process that provides common functionality for various upload methods to AEM.
  */
 class AEMBinaryUpload {
     /**
@@ -68,20 +52,16 @@ class AEMBinaryUpload {
      * @returns {*} Transfer process options.
      */
     getTransferOptions() {
-        const { requestOptions = {} } = this.getOptions();
-        return {
-            retryMaxCount: MAX_RETRY_COUNT,
-            requestOptions
-        };
+        throw new Error("getTransferOptions() must be implemented.");
     }
 
     /**
      * Creates the transfer controller that should be used to transfer the files
      * to AEM.
-     * @returns {TransferController} A transfer controller.
+     * @returns {import('../controller/transfercontroller').TransferController} A transfer controller.
      */
     createTransferController() {
-        return new TransferController();
+        throw new Error("createTransferControll() must be implemented.");
     }
 
     /**
@@ -121,17 +101,9 @@ class AEMBinaryUpload {
      * @returns {import('../asset/transferasset').TransferAssetOptions} Transfer options
      *  to use for the file.
      */
+    // eslint-disable-next-line no-unused-vars
     getTransferAssetOptions(uploadFile) {
-        return {
-            acceptRanges: true,
-            metadata: new AssetMetadata(uploadFile.filePath, undefined, uploadFile.fileSize),
-            nameConflictPolicy: new NameConflictPolicy({
-                createVersion: uploadFile.createVersion,
-                versionLabel: uploadFile.versionLabel,
-                versionComment: uploadFile.versionComment,
-                replace: uploadFile.replace
-            })
-        };
+        throw new Error("getTransferAssetOptions() must be implemented.");
     }
 
     /**
@@ -140,7 +112,7 @@ class AEMBinaryUpload {
      * @returns {number} Preferred part size, or undefined if none specified.
      */
     getPreferredPartSize() {
-        return this.getOptions().preferredPartSize;
+        throw new Error("getPreferredPartSize() must be implemented.");
     }
 
     /**
@@ -148,8 +120,7 @@ class AEMBinaryUpload {
      * @returns {number} Number of max uploads.
      */
     getMaxConcurrent() {
-        const options = this.getOptions();
-        return (options.concurrent && options.maxConcurrent) || 1;
+        throw new Error("getMaxConcurrent() must be implemented.");
     }
 
     /**
@@ -160,19 +131,9 @@ class AEMBinaryUpload {
      * @returns {Array<import('../generator/function')>} Steps to execute in a
      *  pipeline.
      */
+    // eslint-disable-next-line no-unused-vars
     getPipelineSteps(randomFileAccess) {
-        const transferOptions = this.getTransferOptions();
-        return [
-            new FailUnsupportedAssets(),
-            new MapConcurrent(new AEMInitiateUpload(transferOptions), { maxBatchLength: 100 }),
-            new CreateTransferParts({ preferredPartSize: this.getPreferredPartSize() }),
-            new MapConcurrent(new Transfer(randomFileAccess, transferOptions), {
-                maxConcurrent: this.getMaxConcurrent()
-            }),
-            new JoinTransferParts,
-            new CloseFiles(randomFileAccess),
-            new MapConcurrent(new AEMCompleteUpload(transferOptions)),
-        ];
+        throw new Error("getPipelineSteps() must be implemented.");
     }
 
     /**
@@ -181,7 +142,7 @@ class AEMBinaryUpload {
      * @returns {string} An event name.
      */
     getFileStartEventName() {
-        return TransferEvents.AEM_INITIATE_UPLOAD;
+        throw new Error("getFileStartEventName() must be implemented.");
     }
 
     /**
@@ -190,65 +151,7 @@ class AEMBinaryUpload {
      * @returns {string} An event name.
      */
     getFileEndEventName() {
-        return TransferEvents.AFTER_AEM_COMPLETE_UPLOAD;
-    }
-
-    /**
-     * Determines whether the AEM isntance that is the target of the upload has direct
-     * binary access enabled. This will help decide which upload algorithm to use.
-     * @param {import('typedefs').AEMUploadOptions} options Options being used to drive the upload.
-     * @returns {Promise<booleam>} True if direct binary access is enabled, false
-     *  otherwise.
-     */
-    static async isDirectBinaryAccessEnabled(options) {
-        return new Promise((res, rej) => {
-            const controller = new TransferController();
-            controller.on(TransferEvents.AFTER_AEM_INITIATE_UPLOAD, () => {
-                // this event means that the initiate was successful - direct
-                // binary access is support
-                res(true);
-            });
-            controller.on(TransferEvents.ERROR, (transferEvent) => {
-                try {
-                    // ther was an error in the initiate, check to see if the error
-                    // indicates direct binary access isn't enabled
-                    if (transferEvent.props.firstError) {
-                        if (transferEvent.error.code === ErrorCodes.NOT_SUPPORTED) {
-                            res(false);
-                            return;
-                        }
-                    }
-                    // default to true to avoid swallowing errors that downstream processes
-                    // may handle better
-                    res(true);
-                } catch (e) {
-                    rej(e);
-                }
-            });
-            try {
-                const binaryUpload = new AEMBinaryUpload(options);
-                const transferAssets = binaryUpload.generateAEMUploadTransferRecords();
-
-                const assets = [];
-                // use the first asset in the list of files to submit an initiate upload call,
-                // which can help determine whether direct binary access is enabled.
-                for (const transferAsset of transferAssets) {
-                    assets.push(transferAsset);
-                    break;
-                }
-                if (assets.length > 0) {
-                    const initiateUpload = new AEMInitiateUpload(binaryUpload.getTransferOptions());
-                    const result = initiateUpload.execute(assets, controller);
-                    result.next();
-                } else {
-                    // if no assets were provided, continue as if direct binary access is
-                    // enabled and rely on downstream processes to handle the error
-                    res(true);
-                }
-            } catch (e) {
-                rej(e);
-            }
-        });
+        throw new Error("getFileEndEventName() must be implemented.");
     }
 }
