+++ lib/aem/aemupload.js
@@ -20,7 +20,7 @@ const { RandomFileAccess } = require("../randomfileaccess");
 const EventEmitter = require("events");
 const UploadError = require("../block/upload-error");
 const { FilterFailedAssets } = require("../functions/filterfailedassets");
-const { AEMBinaryUpload } = require("./aembinaryupload");
+const { DirectBinaryUpload } = require("./directbinaryupload");
 const { CreateAssetServletUpload } = require("./createassetservletupload");
 
 class AEMUpload extends EventEmitter {
@@ -30,8 +30,8 @@ class AEMUpload extends EventEmitter {
      * @param {import('typedefs').AEMUploadOptions} options AEM upload options
      */
     async uploadFiles(options) {
-        const isDirectBinaryEnabled = await AEMBinaryUpload.isDirectBinaryAccessEnabled(options);
-        const binaryUpload = isDirectBinaryEnabled ? new AEMBinaryUpload(options) : new CreateAssetServletUpload(options);
+        const isDirectBinaryEnabled = await DirectBinaryUpload.isDirectBinaryAccessEnabled(options);
+        const binaryUpload = isDirectBinaryEnabled ? new DirectBinaryUpload(options) : new CreateAssetServletUpload(options);
 
         const controller = binaryUpload.createTransferController();
         controller.on(binaryUpload.getFileStartEventName(), transferEvent => {
