+++ lib/aem/createassetservletupload.js
@@ -34,6 +34,7 @@ const { HTTP } = require("../constants");
 const DEFAULT_PART_SIZE = 1024 * 1024 * 10; // default 10MB parts
 const DEFAULT_URL_COUNT = 1;
 const MAX_CONCURRENT = 1;
+const MAX_RETRY_COUNT = 5;
 const CREATE_ASSET_SERVLET_SELECTOR = ".createasset.html";
 
 /**
@@ -47,10 +48,10 @@ class CreateAssetServletUpload extends AEMBinaryUpload {
      * @returns {*} Transfer process options.
      */
     getTransferOptions() {
-        // overridden to ensure that parts are POSTed to the create asset
-        // servlet
+        const { requestOptions = {} } = this.getOptions();
         return {
-            ...super.getTransferOptions(),
+            retryMaxCount: MAX_RETRY_COUNT,
+            requestOptions,
             method: HTTP.METHOD.POST
         };
     }
@@ -98,7 +99,7 @@ class CreateAssetServletUpload extends AEMBinaryUpload {
      * @returns {number} Preferred part size, or undefined if none specified.
      */
     getPreferredPartSize() {
-        const defaultSize = super.getPreferredPartSize();
+        const defaultSize = this.getOptions().preferredPartSize;
         // overridden to have a default part size of 10MB for the create
         // asset servlet
         return defaultSize || DEFAULT_PART_SIZE;
